requirements.txt
# Core LLM + OpenAI
openai>=1.0.0

# LangChain ecosystem (LCEL + Runnables)
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0

# Tavily search tool
tavily-python>=0.3.0

# Web scraping
beautifulsoup4>=4.12.3
requests>=2.31.0
lxml>=5.0.0

# Async support (important for agents + pipelines)
aiohttp>=3.9.0
httpx>=0.27.0

# Data handling
pydantic>=2.7.0

# Environment variables
python-dotenv>=1.0.1

# Logging / debugging
rich>=13.7.0 

# Optional: vector store (if you expand later)
faiss-cpu>=1.8.0

# Optional: token counting
tiktoken>=0.7.0